--- .eslintrc
@@ -1,28 +1,0 @@
-{
-  "root": true,
-  "env": {
-    "node": true
-  },
-  "extends": ["plugin:vue/essential", "eslint:recommended"],
-  "rules": {},
-  "parserOptions": {
-    "parser": "@babel/eslint-parser"
-  },
-  "overrides": [
-    {
-      "files": ["**/__tests__/*.js"],
-      "parserOptions": {
-        "parser": "@babel/eslint-parser",
-        "sourceType": "module"
-      },
-      "env": {
-        "jest": true
-      },
-      "globals": {
-        "mount": false,
-        "shallowMount": false,
-        "createComponentMocks": false
-      }
-    }
-  ]
-}
