+++ .gitignore
@@ -23,3 +23,4 @@ yarn-error.log*
 *.njsproj
 *.sln
 *.sw?
+cypress/videos
