--- babel.config.js
@@ -1,13 +1,0 @@
-module.exports = {
-  presets: ["@vue/app"],
-  overrides: [
-    {
-      include: ["node_modules"],
-      presets: [
-        [
-          "@babel/env"
-        ]
-      ]
-    }
-  ]
-};
