--- cypress.json
@@ -1,3 +1,0 @@
-{
-  "pluginsFile": "tests/e2e/plugins/index.js"
-}
