+++ cypress/e2e/CheckErrors.cy.js
