--- demo/main.js
@@ -1,5 +1,0 @@
-import "./find.polyfill";
-import { createApp } from "vue";
-import App from "./App.vue";
-
-createApp(App).mount("#app");
