+++ demo/main.ts
@@ -1,0 +1,5 @@
+import "./find.polyfill";
+import { createApp } from 'vue'
+import App from './App.vue'
+
+createApp(App).mount('#app')
