--- jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  preset: '@vue/cli-plugin-unit-jest'
-}
