+++ package.json
@@ -3,48 +3,47 @@
   "version": "0.0.0-dev",
   "author": "Bart Ledoux <ledouxb@me.com>",
   "scripts": {
-    "serve": "vue-cli-service serve",
+    "serve": "vite serve",
     "build": "rollup -c",
-    "lint": "vue-cli-service lint",
-    "build:demo": "vue-cli-service build --modern",
-    "start": "vue-cli-service serve",
-    "test:e2e": "vue-cli-service test:e2e",
-    "test:unit": "vue-cli-service test:unit",
+    "build:demo": "vite build",
+    "start": "vite serve",
+    "preview": "vite preview --port 4173",
+    "test:e2e": "start-server-and-test preview http://localhost:4173/ 'cypress run --e2e'",
+    "test:unit": "vitest run",
     "test": "npm run test:unit && npm run test:e2e",
+    "cy": "cypress open --e2e",
     "cy:verify": "cypress verify"
   },
   "dependencies": {
     "@babel/runtime": "^7.17.2",
     "@vue/compiler-core": "^3.2.37",
-    "@vue/compiler-dom": "^3.2.37",
     "@vue/compiler-sfc": "^3.2.37",
     "acorn": "^8.7.0",
     "acorn-walk": "^8.2.0",
     "core-js": "^3.6.5",
     "debounce": "^1.2.0",
     "hash-sum": "^2.0.0",
-    "prismjs": "^1.27.0",
+    "prismjs": "^1.28.0",
     "vue-inbrowser-compiler": "^4.50.0",
     "vue-prism-editor": "^2.0.0-alpha.2"
   },
   "devDependencies": {
+    "@types/jsdom": "^16.2.14",
+    "@types/node": "^16.11.45",
+    "@types/prismjs": "^1.26.0",
+    "@vue/tsconfig": "^0.1.3",
     "@babel/eslint-parser": "^7.17.0",
     "@babel/plugin-transform-runtime": "^7.17.0",
-    "@rollup/plugin-babel": "^5.1.0",
-    "@rollup/plugin-commonjs": "^14.0.0",
-    "@rollup/plugin-node-resolve": "^8.4.0",
-    "@vue/cli-plugin-babel": "^5.0.1",
-    "@vue/cli-plugin-e2e-cypress": "^5.0.1",
-    "@vue/cli-plugin-eslint": "^5.0.1",
-    "@vue/cli-plugin-unit-jest": "^5.0.1",
-    "@vue/cli-service": "^5.0.1",
-    "@vue/test-utils": "^2.0.0-rc.17",
-    "@vue/vue3-jest": "^27.0.0-alpha.1",
+    "@rollup/plugin-babel": "^5.3.1",
+    "@rollup/plugin-commonjs": "^22.0.1",
+    "@rollup/plugin-node-resolve": "^13.3.0",
+    "@uivjs/vue-github-corners": "^1.0.1",
+    "@vitejs/plugin-vue": "^3.0.1",
+    "@vue/babel-preset-app": "^5.0.8",
+    "autoprefixer": "^10.4.8",
+    "cypress": "^10.3.1",
     "cross-env": "^7.0.2",
-    "eslint": "^8.10.0",
-    "eslint-plugin-vue": "^8.5.0",
     "patch-package": "^6.4.7",
-    "raw-loader": "^4.0.2",
     "rollup": "^2.28.1",
     "rollup-plugin-analyzer": "^3.2.3",
     "rollup-plugin-css-only": "^2.1.0",
@@ -53,7 +52,12 @@
     "vue": "^3.2.31",
     "vue-github-corners": "^1.2.3",
     "vue3-datepicker": "^0.1.0",
-    "webpack": "5"
+    "npm-run-all": "^4.1.5",
+    "start-server-and-test": "^1.14.0",
+    "typescript": "~4.7.4",
+    "vite": "^3.0.1",
+    "vitest": "^0.18.1",
+    "vue-tsc": "^0.38.8"
   },
   "peerDependencies": {
     "vue": ">= 3.0.8"
