+++ src/utils/checkTemplate.spec.ts
@@ -1,7 +1,8 @@
-import checkTemplateDummy from "../checkTemplate";
-import defaultAttrAllowList from "../defaultAttrAllowList";
+import {expect, test} from "vitest"
+import checkTemplateDummy from "./checkTemplate";
+import defaultAttrAllowList from "./defaultAttrAllowList";
 
-const checkTemplate = (opts) => checkTemplateDummy(opts, true);
+const checkTemplate = (opts:any) => checkTemplateDummy(opts, true);
 
 test("parse valid template without error with a function", () => {
   expect(() =>
@@ -114,7 +115,7 @@ test("parse a valid arrow event handler properly", () => {
       template:
         '<div><compo @click="( evt ) => { test(evt); }">hello</compo></div>',
       methods: {
-        test(e) {
+        test(e:any) {
           console.log(e.target);
         },
       },
@@ -128,7 +129,7 @@ test("parse a valid standard event handler properly", () => {
       template:
         '<div><compo @click="function( evt ) { test(evt); }">hello</compo></div>',
       methods: {
-        test(e) {
+        test(e:any) {
           console.log(e.target);
         },
       },
