--- tests/e2e/.eslintrc.js
@@ -1,12 +1,0 @@
-module.exports = {
-  plugins: [
-    'cypress'
-  ],
-  env: {
-    mocha: true,
-    'cypress/globals': true
-  },
-  rules: {
-    strict: 'off'
-  }
-}
